+++ InnerEye/ML/Histopathology/utils/heatmap_utils.py
@@ -1,3 +1,8 @@
+#  ------------------------------------------------------------------------------------------
+#  Copyright (c) Microsoft Corporation. All rights reserved.
+#  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
+#  ------------------------------------------------------------------------------------------
+
 from typing import Optional, List
 import matplotlib.pyplot as plt
 import numpy as np