+++ InnerEye/ML/configs/histo_configs/classification/DeepSMILECrck.py
@@ -57,8 +57,6 @@ class DeepSMILECrck(BaseMIL):
             # To mount the dataset instead of downloading in AML, pass --use_dataset_mount in the CLI
             # declared in TrainerParams:
             num_epochs=16,
-            recovery_checkpoint_save_interval=16,
-            recovery_checkpoints_save_last_k=-1,
             # declared in WorkflowParams:
             number_of_cross_validation_splits=5,
             cross_validation_split_index=0,