+++ InnerEye/ML/configs/ssl/CIFAR_SSL_configs.py
@@ -20,7 +20,6 @@ class CIFAR10SimCLR(SSLContainer):
                          ssl_encoder=EncoderName.resnet50,
                          ssl_training_type=SSLTrainingType.SimCLR,
                          random_seed=1,
-                         recovery_checkpoint_save_interval=200,
                          num_epochs=2500,
                          num_workers=6)
 
@@ -38,7 +37,6 @@ class CIFAR10BYOL(SSLContainer):
                          ssl_encoder=EncoderName.resnet50,
                          ssl_training_type=SSLTrainingType.BYOL,
                          random_seed=1,
-                         recovery_checkpoint_save_interval=200,
                          num_epochs=2500,
                          num_workers=6)
 
@@ -55,6 +53,5 @@ class CIFAR10CIFAR100BYOL(SSLContainer):
                          ssl_encoder=EncoderName.resnet50,
                          ssl_training_type=SSLTrainingType.BYOL,
                          random_seed=1,
-                         recovery_checkpoint_save_interval=200,
                          num_epochs=2500,
                          num_workers=6)